requirements.txt
Django
django-basic-authentication-decorator
